Refactor code structure for improved readability and maintainability

+++ apps/frontend/lib/features/retail/pos/presentation/state/parked_carts_notifier.dart
@@ -23,7 +23,7 @@ class ParkedCartsNotifier extends StateNotifier<List<CartState>> {
           ..remoteId = i.productId
           ..name = i.name!
           ..price = i.price,
-        quantity: i.quantity.toInt(),
+        quantity: i.quantity,
         discountAmount: i.discountAmount,
         discountType: i.discountType,
       )).toList(),
